backend/requirements.txt
alembic==1.6.5
blinker==1.4
click==7.1.2
Flask==1.1.4
Flask-Alembic==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
flask-praetorian==1.2.0
Flask-SQLAlchemy==2.5.1
flask-wtf==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==2.0.1
pandas==1.3.3
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
Werkzeug==1.0.1
